got core qa graph working
