+++ backend/danswer/agent_search/core_qa_graph/edges.py
@@ -3,7 +3,7 @@ from typing import Union
 
 from langgraph.types import Send
 
-from danswer.agent_search.base_qa_sub_graph.states import BaseQAState
+from danswer.agent_search.core_qa_graph.states import BaseQAState
 from danswer.agent_search.primary_graph.states import RetrieverState
 from danswer.agent_search.primary_graph.states import VerifierState
 
@@ -18,8 +18,6 @@ def sub_continue_to_verifier(state: BaseQAState) -> Union[Hashable, list[Hashabl
             VerifierState(
                 document=doc,
                 question=state["sub_question_str"],
-                fast_llm=state["fast_llm"],
-                primary_llm=state["primary_llm"],
                 graph_start_time=state["graph_start_time"],
             ),
         )
@@ -30,15 +28,14 @@ def sub_continue_to_verifier(state: BaseQAState) -> Union[Hashable, list[Hashabl
 def sub_continue_to_retrieval(state: BaseQAState) -> Union[Hashable, list[Hashable]]:
     # Routes re-written queries to the (parallel) retrieval steps
     # Notice the 'Send()' API that takes care of the parallelization
+    rewritten_queries = state["sub_question_search_queries"].rewritten_queries
     return [
         Send(
             "sub_custom_retrieve",
             RetrieverState(
                 rewritten_query=query,
-                primary_llm=state["primary_llm"],
-                fast_llm=state["fast_llm"],
                 graph_start_time=state["graph_start_time"],
             ),
         )
-        for query in state["sub_question_search_queries"]
+        for query in rewritten_queries
     ]
