+++ backend/danswer/agent_search/core_qa_graph/states.py
@@ -7,7 +7,8 @@ from typing import TypedDict
 from langchain_core.messages import BaseMessage
 from langgraph.graph.message import add_messages
 
-from danswer.chat.models import DanswerContext
+from danswer.agent_search.shared_graph_utils.models import RewrittenQueries
+from danswer.context.search.models import InferenceSection
 from danswer.llm.interfaces import LLM
 
 
@@ -18,8 +19,13 @@ class SubQuestionRetrieverState(TypedDict):
 
 class SubQuestionVerifierState(TypedDict):
     # The state for the parallel verification step.  Each node execution need to see only one question/doc pair
-    sub_question_document: DanswerContext
+    sub_question_document: InferenceSection
     sub_question: str
+
+
+class CoreQAInputState(TypedDict):
+    sub_question_str: str
+    original_question: str
 
 
 class BaseQAState(TypedDict):
@@ -30,17 +36,19 @@ class BaseQAState(TypedDict):
     sub_query_start_time: datetime
     sub_question_rewritten_queries: list[str]
     sub_question_str: str
-    sub_question_search_queries: list[str]
+    sub_question_search_queries: RewrittenQueries
     sub_question_nr: int
-    sub_question_base_retrieval_docs: Annotated[Sequence[DanswerContext], operator.add]
+    sub_question_base_retrieval_docs: Annotated[
+        Sequence[InferenceSection], operator.add
+    ]
     sub_question_deduped_retrieval_docs: Annotated[
-        Sequence[DanswerContext], operator.add
+        Sequence[InferenceSection], operator.add
     ]
     sub_question_verified_retrieval_docs: Annotated[
-        Sequence[DanswerContext], operator.add
+        Sequence[InferenceSection], operator.add
     ]
     sub_question_reranked_retrieval_docs: Annotated[
-        Sequence[DanswerContext], operator.add
+        Sequence[InferenceSection], operator.add
     ]
     sub_question_top_chunks: Annotated[Sequence[dict], operator.add]
     sub_question_answer: str
@@ -63,15 +71,17 @@ class BaseQAOutputState(TypedDict):
     sub_qas: Annotated[Sequence[dict], operator.add]
     # Answers sent back to core
     initial_sub_qas: Annotated[Sequence[dict], operator.add]
-    sub_question_base_retrieval_docs: Annotated[Sequence[DanswerContext], operator.add]
+    sub_question_base_retrieval_docs: Annotated[
+        Sequence[InferenceSection], operator.add
+    ]
     sub_question_deduped_retrieval_docs: Annotated[
-        Sequence[DanswerContext], operator.add
+        Sequence[InferenceSection], operator.add
     ]
     sub_question_verified_retrieval_docs: Annotated[
-        Sequence[DanswerContext], operator.add
+        Sequence[InferenceSection], operator.add
     ]
     sub_question_reranked_retrieval_docs: Annotated[
-        Sequence[DanswerContext], operator.add
+        Sequence[InferenceSection], operator.add
     ]
     sub_question_top_chunks: Annotated[Sequence[dict], operator.add]
     sub_question_answer: str
