+++ backend/danswer/agent_search/primary_graph/states.py
@@ -7,8 +7,8 @@ from typing import TypedDict
 from langchain_core.messages import BaseMessage
 from langgraph.graph.message import add_messages
 
-from danswer.chat.models import DanswerContext
-from danswer.llm.interfaces import LLM
+from danswer.agent_search.shared_graph_utils.models import RewrittenQueries
+from danswer.context.search.models import InferenceSection
 
 
 class QAState(TypedDict):
@@ -18,7 +18,7 @@ class QAState(TypedDict):
     # start time for parallel initial sub-questionn thread
     sub_query_start_time: datetime
     log_messages: Annotated[Sequence[BaseMessage], add_messages]
-    rewritten_queries: list[str]
+    rewritten_queries: RewrittenQueries
     sub_questions: list[dict]
     initial_sub_questions: list[dict]
     ranked_subquestion_ids: list[int]
@@ -28,13 +28,13 @@ class QAState(TypedDict):
     sub_qas: Annotated[Sequence[dict], operator.add]
     initial_sub_qas: Annotated[Sequence[dict], operator.add]
     checked_sub_qas: Annotated[Sequence[dict], operator.add]
-    base_retrieval_docs: Annotated[Sequence[DanswerContext], operator.add]
-    deduped_retrieval_docs: Annotated[Sequence[DanswerContext], operator.add]
-    reranked_retrieval_docs: Annotated[Sequence[DanswerContext], operator.add]
+    base_retrieval_docs: Annotated[Sequence[InferenceSection], operator.add]
+    deduped_retrieval_docs: Annotated[Sequence[InferenceSection], operator.add]
+    reranked_retrieval_docs: Annotated[Sequence[InferenceSection], operator.add]
     retrieved_entities_relationships: dict
     questions_context: list[dict]
     qa_level: int
-    top_chunks: list[DanswerContext]
+    top_chunks: list[InferenceSection]
     sub_question_top_chunks: Annotated[Sequence[dict], operator.add]
     num_new_question_iterations: int
     core_answer_dynamic_context: str
@@ -42,8 +42,6 @@ class QAState(TypedDict):
     initial_base_answer: str
     base_answer: str
     deep_answer: str
-    primary_llm: LLM
-    fast_llm: LLM
 
 
 class QAOuputState(TypedDict):
@@ -54,9 +52,9 @@ class QAOuputState(TypedDict):
     sub_qas: Annotated[Sequence[dict], operator.add]
     initial_sub_qas: Annotated[Sequence[dict], operator.add]
     checked_sub_qas: Annotated[Sequence[dict], operator.add]
-    reranked_retrieval_docs: Annotated[Sequence[DanswerContext], operator.add]
+    reranked_retrieval_docs: Annotated[Sequence[InferenceSection], operator.add]
     retrieved_entities_relationships: dict
-    top_chunks: list[DanswerContext]
+    top_chunks: list[InferenceSection]
     sub_question_top_chunks: Annotated[Sequence[dict], operator.add]
     base_answer: str
     deep_answer: str
@@ -65,15 +63,11 @@ class QAOuputState(TypedDict):
 class RetrieverState(TypedDict):
     # The state for the parallel Retrievers. They each need to see only one query
     rewritten_query: str
-    primary_llm: LLM
-    fast_llm: LLM
     graph_start_time: datetime
 
 
 class VerifierState(TypedDict):
     # The state for the parallel verification step.  Each node execution need to see only one question/doc pair
-    document: DanswerContext
+    document: InferenceSection
     question: str
-    primary_llm: LLM
-    fast_llm: LLM
     graph_start_time: datetime
